refactor: update path handling to use lrncore::path functions
