+++ src/projects/list/mod.rs
@@ -79,7 +79,7 @@ pub fn create_repo_or_not(tech: &str) {
 }
 
 pub fn add_project_to_list(tech: &str) {
-    let current_dir = utils::get_current_path();
+    let current_dir = lrncore::path::get_current_path();
     let app_name = current_dir.split("/").last().unwrap();
     let mut repository_user_input: String;
     repository_user_input = utils::prompt_message(
@@ -125,7 +125,7 @@ pub fn add_project_to_list(tech: &str) {
 }
 
 fn create_repo_add_to_list(tech: &str) {
-    let current_dir = utils::get_current_path();
+    let current_dir = lrncore::path::get_current_path();
     let app_name = current_dir.split("/").last().unwrap();
     let choice = vec_of_strings!["public", "private", "internal"];
     let repository_visibility: String =
